requirement.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
billiard==3.6.4.0
cached-property==1.2.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
Django==4.1.4
django-cleanup==7.0.0
django-filter==22.1
django-redis==5.2.0
django-templated-mail==1.1.1
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.25.1
exceptiongroup==1.1.0
flower==1.2.0
Faker==18.11.1
gunicorn==20.1.0
humanize==4.6.0
inflection==0.5.1
iniconfig==1.1.1
jsonschema==4.17.3
kombu==5.2.4
Markdown==3.4.1
packaging==22.0
Pillow==9.4.0
pluggy==1.0.0
prometheus-client==0.16.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
PyJWT==2.6.0
pyrsistent==0.19.3
pytest==7.4.0
pytest-django==4.5.2
pytest-env==0.8.1
python-decouple==3.6
pytz==2022.7
PyYAML==6.0
redis==4.4.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tornado==6.2
tzdata==2022.7
uritemplate==4.1.1
vine==5.0.0
wcwidth==0.2.5
